+++ tests/Command/JSHintCommandTest.php
@@ -58,13 +58,12 @@ class JSHintCommandTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(0, $commandTester->getStatusCode());
     }
 
-    /**
-     * @expectedException \RuntimeException
-     */
     public function testExecuteNoFiles()
     {
         // Just random directory with no JS files.
-        $this->executeCommand($this->pluginDir.'/tests/behat');
+        $commandTester = $this->executeCommand($this->pluginDir.'/tests/behat');
+        $this->assertEquals(0, $commandTester->getStatusCode());
+        $this->assertRegExp('/No relevant files found to process, free pass!/', $commandTester->getDisplay());
     }
 
     /**